add abstract component & replace utils functions
