+++ src/components/abstract-component.js
@@ -1,17 +1,14 @@
 import {createElement} from '../utils/render.js';
 
-const createNoTripPointsMarkup = () => {
-  return (
-    `<p class="trip-events__msg">Click New Event to create your first point</p>`
-  );
-};
-
-export default class NoTripPoints {
+export default class AbstractComponent {
   constructor() {
+    if (new.target === AbstractComponent) {
+      throw new Error(`Can't instantiate AbstractComponent, only concrete one.`);
+    }
     this._element = null;
   }
   getTemplate() {
-    return createNoTripPointsMarkup();
+    throw new Error(`Abstract method not implemented: getTemplate`);
   }
   getElement() {
     if (!this._element) {
