+++ src/components/filter.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import {createElement} from '../utils/render.js';
 
 const filterOptions = [`Everything`, `Future`, `Past`];
