+++ src/components/sorting.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import {createElement} from '../utils/render.js';
 
 const sortingList = [`Event`, `Time`, `Price`];
