+++ src/components/trip-edit.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import {createElement} from '../utils/render.js';
 
 let moment = require(`moment`);
 moment().format();
