+++ src/components/trip-offer.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import {createElement} from '../utils/render.js';
 
 const sequence = (start = 1) => {
   let callNumber = start;
